pyproject.toml
[tool.poetry]
name = "blogfull"
version = "0.1.0"
description = ""
authors = ["ptostopank <52743103+ptostopank@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
djoser = "^2.1.0"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
